refactor(workspace): collapse stale markdown indirection
