+++ packages/workspace/src/client/daemon-actions.ts
@@ -52,13 +52,14 @@ export type DaemonActions<TActions> = Simplify<{
 }>;
 
 /**
- * Flat proxy rooted at the action registry. Property access returns a
- * callable for that key; calling it dispatches `client.run` with the joined
- * `${route}.${key}` path. `then` is masked at the root (one place, not at
- * every level) so an accidental `await` on the workspace handle does not
- * turn it into a thenable.
+ * Compose the daemon action facade. Generic `TActions` is the in-process
+ * `ActionRegistry`; `DaemonActions<TActions>` rewrites each entry to the
+ * daemon `/run` result shape.
  */
-function buildDaemonActionProxy(client: DaemonClient, route: string): unknown {
+export function buildDaemonActions<TActions extends ActionRegistry>(
+	client: DaemonClient,
+	route: string,
+): DaemonActions<TActions> {
 	return new Proxy({} as Record<string, unknown>, {
 		get(_target, prop) {
 			if (typeof prop !== 'string') return undefined;
@@ -70,17 +71,5 @@ function buildDaemonActionProxy(client: DaemonClient, route: string): unknown {
 					waitMs: options?.waitMs ?? DEFAULT_RUN_WAIT_MS,
 				});
 		},
-	});
-}
-
-/**
- * Compose the daemon action facade. Generic `TActions` is the in-process
- * `ActionRegistry`; `DaemonActions<TActions>` rewrites each entry to the
- * daemon `/run` result shape.
- */
-export function buildDaemonActions<TActions extends ActionRegistry>(
-	client: DaemonClient,
-	route: string,
-): DaemonActions<TActions> {
-	return buildDaemonActionProxy(client, route) as DaemonActions<TActions>;
+	}) as DaemonActions<TActions>;
 }
