+++ packages/workspace/src/document/materializer/markdown/index.ts
@@ -1,4 +1,14 @@
-export { assembleMarkdown, type SerializeResult } from './markdown.js';
+export {
+	assembleMarkdown,
+	type SerializeResult,
+} from '../../markdown/markdown.js';
+export { parseMarkdownFile } from '../../markdown/parse-markdown-file.js';
+export { prepareMarkdownFiles } from '../../markdown/prepare-markdown-files.js';
+export {
+	slugFilename,
+	toIdFilename,
+	toSlugFilename,
+} from '../../markdown/serializers.js';
 export {
 	attachMarkdownMaterializer,
 	type MarkdownShape,
@@ -6,6 +16,3 @@ export {
 	type PushEvent,
 	type PushResult,
 } from './materializer.js';
-export { parseMarkdownFile } from './parse-markdown-file.js';
-export { prepareMarkdownFiles } from './prepare-markdown-files.js';
-export { slugFilename, toIdFilename, toSlugFilename } from './serializers.js';
