+++ packages/workspace/src/document/materializer/markdown/materializer.test.ts
@@ -26,11 +26,11 @@ import {
 	createDisposableCache,
 	defineTable,
 } from '../../../index.js';
+import { parseMarkdownFile } from '../../markdown/parse-markdown-file.js';
 import {
 	attachMarkdownMaterializer,
 	type MarkdownShape,
 } from './materializer.js';
-import { parseMarkdownFile } from './parse-markdown-file.js';
 
 // ============================================================================
 // Test Table Definitions
@@ -195,7 +195,7 @@ describe('push', () => {
 
 	test('silently skips tables whose directories do not exist', async () => {
 		const { workspace } = await setup({ tables: (t) => [{ table: t.posts }] });
-		// Don't create the posts directory — it should not exist
+		// Don't create the posts directory. It should not exist.
 		const result = await workspace.materializer.push({});
 
 		expect(result.imported).toBe(0);
@@ -207,7 +207,7 @@ describe('push', () => {
 
 	test('emits error event when frontmatter fails schema validation', async () => {
 		const { workspace } = await setup({ tables: (t) => [{ table: t.posts }] });
-		// Valid frontmatter structure but wrong type — title must be a string,
+		// Valid frontmatter structure but wrong type: title must be a string,
 		// here it's a number. `fromMarkdown` happily returns it; `table.parse()`
 		// catches the schema violation.
 		await writeTestFile(
@@ -575,7 +575,7 @@ describe('rebuild', () => {
 		expect(result.deleted).toBe(1); // p1.md
 		expect(result.written).toBe(1); // p1 re-written
 
-		// notes/ is untouched — orphan still there
+		// notes/ is untouched; orphan still there
 		const notesEntries = await listTestDir('notes');
 		expect(notesEntries).toContain('orphan.md');
 
@@ -591,7 +591,7 @@ describe('rebuild', () => {
 		workspace[Symbol.dispose]();
 	});
 
-	test('is idempotent — rebuild twice produces identical filesystem state', async () => {
+	test('is idempotent: rebuild twice produces identical filesystem state', async () => {
 		const { workspace } = await setup({ tables: (t) => [{ table: t.posts }] });
 		workspace.tables.posts.set({
 			id: 'p1',
@@ -677,8 +677,8 @@ describe('round-trip', () => {
 		// Verify files on disk have valid frontmatter
 		const p1Content = await readTestFile('posts/p1.md');
 		const p1Parsed = parseMarkdownFile(p1Content);
-		expect(p1Parsed).not.toBeNull();
-		expect(p1Parsed!.frontmatter.title).toBe('Round Trip');
+		if (p1Parsed === null) throw new Error('expected p1.md to parse');
+		expect(p1Parsed.frontmatter.title).toBe('Round Trip');
 
 		// Second workspace: fresh instance, push from the same directory
 		const cache2 = createDisposableCache((id: string) => {
@@ -715,7 +715,7 @@ describe('round-trip', () => {
 		workspace2[Symbol.dispose]();
 	});
 
-	test('fromMarkdown(toMarkdown(row)) preserves row — MarkdownShape round-trip', async () => {
+	test('fromMarkdown(toMarkdown(row)) preserves row: MarkdownShape round-trip', async () => {
 		// Explicit toMarkdown / fromMarkdown pair over the shared MarkdownShape
 		// type, so the compiler guarantees one is the inverse of the other.
 		const toMarkdownFn = (row: {
@@ -754,8 +754,8 @@ describe('round-trip', () => {
 		// And verify the end-to-end disk round trip.
 		const disk = await readTestFile('notes/n1.md');
 		const parsed = parseMarkdownFile(disk);
-		expect(parsed).not.toBeNull();
-		expect(fromMarkdownFn(parsed!)).toEqual(original);
+		if (parsed === null) throw new Error('expected notes/n1.md to parse');
+		expect(fromMarkdownFn(parsed)).toEqual(original);
 
 		workspace[Symbol.dispose]();
 	});
@@ -763,7 +763,7 @@ describe('round-trip', () => {
 	test('inline field-to-body pair round-trips over MarkdownShape', async () => {
 		// Most real apps store body content in a separate Y.Doc (via
 		// createDisposableCache). This test covers the simpler case where body IS a
-		// row field — `notes.body` here. Inline callbacks keep the intent
+		// row field: `notes.body` here. Inline callbacks keep the intent
 		// at the call site; no helper abstracts the destructure.
 		const { workspace } = await setup({
 			tables: (t) =>
@@ -792,10 +792,10 @@ describe('round-trip', () => {
 
 		const disk = await readTestFile('notes/n1.md');
 		const parsed = parseMarkdownFile(disk);
-		expect(parsed).not.toBeNull();
+		if (parsed === null) throw new Error('expected notes/n1.md to parse');
 		// Body ended up in the markdown body section, not frontmatter.
-		expect(parsed!.body).toBe('Body content here');
-		expect(parsed!.frontmatter.body).toBeUndefined();
+		expect(parsed.body).toBe('Body content here');
+		expect(parsed.frontmatter.body).toBeUndefined();
 
 		workspace[Symbol.dispose]();
 	});
