+++ packages/workspace/src/document/materializer/markdown/materializer.ts
@@ -13,20 +13,16 @@ import { defineMutation } from '../../../shared/actions.js';
 import type { MaybePromise } from '../../../shared/types.js';
 import type { Kv } from '../../attach-kv.js';
 import type { BaseRow, Table, TableParseError } from '../../attach-table.js';
-import type { SerializeResult } from './markdown.js';
-import { assembleMarkdown } from './markdown.js';
-import { parseMarkdownFile } from './parse-markdown-file.js';
+import {
+	assembleMarkdown,
+	type SerializeResult,
+} from '../../markdown/markdown.js';
+import { parseMarkdownFile } from '../../markdown/parse-markdown-file.js';
 
 // ════════════════════════════════════════════════════════════════════════════
 // PUSH ERROR + EVENT TYPES
 // ════════════════════════════════════════════════════════════════════════════
 
-/**
- * Errors produced during `push` that aren't already covered by
- * `TableParseError`. Filename / tableName provenance lives on the
- * emitted `PushEvent`, not inside the error — the error knows its
- * layer, the event carries external context.
- */
 /**
  * Errors produced by the background write-observer (table row → .md file,
  * KV state → serialized file). These run inside `.catch(...)` of a detached
@@ -72,7 +68,7 @@ export type MaterializerPushError = InferErrors<typeof MaterializerPushError>;
  * - **`skipped`**: the file couldn't be parsed as markdown-with-frontmatter
  *   (no `---` delimiters, empty delimiters, or frontmatter that doesn't
  *   decode to an object). The three cases collapse at the parser boundary
- *   into a single "not a note" decision — no discriminator needed.
+ *   into a single "not a note" decision; no discriminator needed.
  * - **`error`**: something failed. `error.name` discriminates between
  *   `ReadFailed` / `FromMarkdownCallbackFailed` (materializer errors) and
  *   `ValidationFailed` / `MigrationFailed` / `AsyncSchemaNotSupported`
@@ -80,7 +76,7 @@ export type MaterializerPushError = InferErrors<typeof MaterializerPushError>;
  *
  * `path` is the relative path from the materializer's base `dir` (e.g.,
  * `"posts/hello.md"` for a file under `config.dir: 'posts'`). Not the
- * bare filename — two tables writing the same filename would be
+ * bare filename: two tables writing the same filename would be
  * indistinguishable otherwise.
  */
 export type PushEvent =
@@ -108,7 +104,7 @@ type AnyTable = Table<any>;
 
 /**
  * Symmetric shape of a parsed markdown file. `toMarkdown` produces it,
- * `fromMarkdown` consumes it — `Parameters<fromMarkdown>[0]` ≡ `ReturnType<toMarkdown>`.
+ * `fromMarkdown` consumes it: `Parameters<fromMarkdown>[0]` ≡ `ReturnType<toMarkdown>`.
  */
 export type MarkdownShape = {
 	frontmatter: Record<string, unknown>;
@@ -133,7 +129,7 @@ type KvConfig = {
 
 type RegisteredTable = {
 	table: AnyTable;
-	// biome-ignore lint/suspicious/noExplicitAny: internal storage — variance across heterogeneous row types
+	// biome-ignore lint/suspicious/noExplicitAny: internal storage, variance across heterogeneous row types
 	config: TableConfig<any>;
 	unsubscribe?: () => void;
 };
@@ -211,13 +207,13 @@ async function writeMarkdownFile(
  * in a single KV mirror. Nothing materializes by default.
  *
  * Exposes three mutations:
- * - `push`    — disk → workspace. Import .md files as rows (additive).
- * - `pull`    — workspace → disk. Write every row as .md file (additive).
- * - `rebuild` — workspace → disk, destructive. Clear output dir then rewrite
+ * - `push`: disk → workspace. Import .md files as rows (additive).
+ * - `pull`: workspace → disk. Write every row as .md file (additive).
+ * - `rebuild`: workspace → disk, destructive. Clear output dir then rewrite
  *   all rows. Use for orphan cleanup or after config changes.
  *   Matches the sqlite materializer's `rebuild` for cross-materializer parity.
  *
- * Teardown is hooked to the ydoc via `ydoc.once('destroy', ...)` — callers
+ * Teardown is hooked to the ydoc via `ydoc.once('destroy', ...)`; callers
  * never call a dispose method; destroying the ydoc cascades.
  *
  * @example
@@ -233,7 +229,7 @@ async function writeMarkdownFile(
  * })
  *   .table(tables.posts, {
  *     filename: slugFilename('title'),
- *     // Inline toMarkdown / fromMarkdown callbacks when needed —
+ *     // Inline toMarkdown / fromMarkdown callbacks when needed:
  *     // most real tables split metadata (on the row) from body
  *     // content (in a separate content-doc via createDisposableCache).
  *   })
@@ -267,7 +263,7 @@ export function attachMarkdownMaterializer(
 	let isDisposed = false;
 	/**
 	 * Closed once `initialize()` commits (past `await waitFor`). Any `.table()`
-	 * / `.kv()` call after this throws — the materializer is past the point
+	 * / `.kv()` call after this throws: the materializer is past the point
 	 * where late registrations would be picked up for initial flush.
 	 */
 	let isRegistrationOpen = true;
@@ -292,7 +288,7 @@ export function attachMarkdownMaterializer(
 			filenames.set(row.id, filename);
 		}
 
-		// Sequential writes inside the observer avoid rename races — a parallel
+		// Sequential writes inside the observer avoid rename races; a parallel
 		// approach (Promise.allSettled) could delete a file another write needs.
 		return table.observe((changedIds) => {
 			void (async () => {
@@ -412,7 +408,7 @@ export function attachMarkdownMaterializer(
 			for (const filename of files) {
 				if (!filename.endsWith('.md')) continue;
 
-				// Relative to the materializer's base dir — disambiguates two
+				// Relative to the materializer's base dir; disambiguates two
 				// tables writing files with the same name.
 				const path = join(subdir, filename);
 
@@ -433,7 +429,7 @@ export function attachMarkdownMaterializer(
 					continue;
 				}
 
-				// 3. Run user's fromMarkdown (or default) — capture throws as errors
+				// 3. Run user's fromMarkdown (or default), capture throws as errors
 				const fromMarkdown: (p: MarkdownShape) => MaybePromise<BaseRow> =
 					entry.config.fromMarkdown ?? defaultFromMarkdown;
 				const { data: row, error: callbackError } = await tryAsync({
@@ -494,7 +490,7 @@ export function attachMarkdownMaterializer(
 
 		if (tableName !== undefined && targets.length === 0) {
 			throw new Error(
-				`Cannot rebuild "${tableName}" — not in the materialized table set.`,
+				`Cannot rebuild "${tableName}": not in the materialized table set.`,
 			);
 		}
 
@@ -510,7 +506,7 @@ export function attachMarkdownMaterializer(
 					deleted++;
 				}
 			} catch {
-				// Directory doesn't exist yet — fine.
+				// Directory doesn't exist yet. Fine.
 			}
 
 			await mkdir(directory, { recursive: true });
@@ -574,7 +570,7 @@ export function attachMarkdownMaterializer(
 		rebuild: defineMutation({
 			title: 'Rebuild markdown files',
 			description:
-				'Delete existing .md files in registered table directories and re-serialize all valid rows. Destructive — removes orphan files left by deleted rows or stale configs.',
+				'Delete existing .md files in registered table directories and re-serialize all valid rows. Destructive: removes orphan files left by deleted rows or stale configs.',
 			input: Type.Object({ table: Type.Optional(Type.String()) }),
 			handler: ({ table }) => rebuildMarkdownFiles(table),
 		}),
