--- packages/workspace/src/document/materializer/markdown/prepare-markdown-files.ts
@@ -1,8 +1,0 @@
-/**
- * Compatibility export for the markdown materializer package path.
- *
- * Markdown file preparation has one implementation in `document/markdown`.
- * This bridge keeps the materializer API stable while the shared tests cover
- * the only behavior source.
- */
-export { prepareMarkdownFiles } from '../../markdown/prepare-markdown-files.js';
